refactor(board-domain): use base-package as `nettee.board.domain`
