+++ services/board/api/domain/src/main/java/nettee/board/domain/type/BoardStatus.java
@@ -1,4 +1,4 @@
-package nettee.board.type;
+package nettee.board.domain.type;
 
 import nettee.common.util.EnumUtil;
