+++ services/board/application/src/main/java/nettee/board/port/BoardCommandRepositoryPort.java
@@ -1,6 +1,6 @@
 package nettee.board.port;
 
-import nettee.board.Board;
+import nettee.board.domain.Board;
 
 public interface BoardCommandRepositoryPort {
     Board save(Board board);
